+++ sandbox/src/game.ts
@@ -894,13 +894,37 @@ game.input.keyboard.on('down', (keyDown?: ex.KeyEvent) => {
     });
     game.add(a);
   } else if (keyDown.key === ex.Keys.U) {
-    game.goToScene('label', { destinationIn: new ex.Slide({ duration: 1000, slideDirection: 'up' }) });
+    game.goToScene('label', {
+      destinationIn: new ex.Slide({
+        duration: 1000,
+        easingFunction: ex.EasingFunctions.EaseInOutCubic,
+        slideDirection: 'up'
+      })
+    });
   } else if (keyDown.key === ex.Keys.D) {
-    game.goToScene('label', { destinationIn: new ex.Slide({ duration: 1000, slideDirection: 'down' }) });
+    game.goToScene('label', {
+      destinationIn: new ex.Slide({
+        duration: 1000,
+        easingFunction: ex.EasingFunctions.EaseInOutCubic,
+        slideDirection: 'down'
+      })
+    });
   } else if (keyDown.key === ex.Keys.L) {
-    game.goToScene('label', { destinationIn: new ex.Slide({ duration: 1000, slideDirection: 'left' }) });
+    game.goToScene('label', {
+      destinationIn: new ex.Slide({
+        duration: 1000,
+        easingFunction: ex.EasingFunctions.EaseInOutCubic,
+        slideDirection: 'left'
+      })
+    });
   } else if (keyDown.key === ex.Keys.R) {
-    game.goToScene('label', { destinationIn: new ex.Slide({ duration: 1000, slideDirection: 'right' }) });
+    game.goToScene('label', {
+      destinationIn: new ex.Slide({
+        duration: 1000,
+        easingFunction: ex.EasingFunctions.EaseInOutCubic,
+        slideDirection: 'right'
+      })
+    });
   } else if (keyDown.key === ex.Keys.I) {
     game.goToScene('root', {
       destinationIn: new ex.CrossFade({